Extract 'DashCardCardParameterMapperButton'
