+++ frontend/src/metabase/dashboard/components/DashCard/DashCardParameterMapper/DashCardCardParameterMapper.tsx
@@ -5,9 +5,7 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import { isActionDashCard } from "metabase/actions/utils";
-import TippyPopover from "metabase/components/Popover/TippyPopover";
 import { Ellipsified } from "metabase/core/components/Ellipsified";
-import DeprecatedTooltip from "metabase/core/components/Tooltip";
 import CS from "metabase/css/core/index.css";
 import { setParameterMapping } from "metabase/dashboard/actions/parameters";
 import {
@@ -23,7 +21,6 @@ import {
   showVirtualDashCardInfoText,
   isQuestionDashCard,
 } from "metabase/dashboard/utils";
-import ParameterTargetList from "metabase/parameters/components/ParameterTargetList";
 import type { ParameterMappingOption } from "metabase/parameters/utils/mapping-options";
 import { getIsRecentlyAutoConnectedDashcard } from "metabase/redux/undo";
 import { getMetadata } from "metabase/selectors/metadata";
@@ -55,30 +52,12 @@ import {
   Container,
   CardLabel,
   Header,
-  TargetButton,
-  TargetButtonText,
   TextCardDefault,
-  CloseIconButton,
-  ChevrondownIcon,
-  KeyIcon,
   Warning,
 } from "./DashCardCardParameterMapper.styled";
+import { DashCardCardParameterMapperButton } from "./DashCardCardParameterMapperButton";
 import { DisabledNativeCardHelpText } from "./DisabledNativeCardHelpText";
 import { useResetParameterMapping } from "./hooks";
-
-function formatSelected({
-  name,
-  sectionName,
-}: {
-  name: string;
-  sectionName?: string;
-}) {
-  if (sectionName == null) {
-    // for native question variables or field literals we just display the name
-    return name;
-  }
-  return `${sectionName}.${name}`;
-}
 
 const mapStateToProps = (
   state: State,
@@ -134,16 +113,12 @@ export function DashCardCardParameterMapper({
   mappingOptions,
   isRecentlyAutoConnected,
 }: DashcardCardParameterMapperProps) {
-  const [isDropdownVisible, setIsDropdownVisible] = useState(false);
-
-  const hasSeries =
-    isQuestionDashCard(dashcard) &&
-    dashcard.series &&
-    dashcard.series.length > 0;
+  const isQuestion = isQuestionDashCard(dashcard);
+  const hasSeries = isQuestion && dashcard.series && dashcard.series.length > 0;
   const isDisabled = mappingOptions.length === 0 || isActionDashCard(dashcard);
   const isVirtual = isVirtualDashCard(dashcard);
   const virtualCardType = getVirtualCardType(dashcard);
-  const isNative = isQuestionDashCard(dashcard) && isNativeDashCard(dashcard);
+  const isNative = isQuestion && isNativeDashCard(dashcard);
   const isTemporalUnit =
     editingParameter != null && isTemporalUnitParameter(editingParameter);
 
@@ -170,85 +145,6 @@ export function DashCardCardParameterMapper({
     question,
     editingParameter ?? undefined,
   );
-
-  const hasPermissionsToMap = useMemo(() => {
-    if (isVirtual) {
-      return true;
-    }
-
-    // virtual or action dashcard
-    if (!isQuestionDashCard(dashcard)) {
-      return true;
-    }
-
-    if (!question || !card.dataset_query) {
-      return false;
-    }
-
-    return question.canRunAdhocQuery();
-  }, [isVirtual, dashcard, card.dataset_query, question]);
-
-  const { buttonVariant, buttonTooltip, buttonText, buttonIcon } =
-    useMemo(() => {
-      if (!hasPermissionsToMap) {
-        return {
-          buttonVariant: "unauthed",
-          buttonTooltip: t`You don’t have permission to see this question’s columns.`,
-          buttonText: null,
-          buttonIcon: <KeyIcon name="key" />,
-        };
-      } else if (isDisabled && !isVirtual) {
-        return {
-          buttonVariant: "disabled",
-          buttonTooltip: t`This card doesn't have any fields or parameters that can be mapped to this parameter type.`,
-          buttonText: t`No valid fields`,
-          buttonIcon: null,
-        };
-      } else if (selectedMappingOption) {
-        return {
-          buttonVariant: "mapped",
-          buttonTooltip: null,
-          buttonText: formatSelected(selectedMappingOption),
-          buttonIcon: (
-            <CloseIconButton
-              role="button"
-              aria-label={t`Disconnect`}
-              onClick={e => {
-                handleChangeTarget(null);
-                e.stopPropagation();
-              }}
-            />
-          ),
-        };
-      } else if (target != null) {
-        return {
-          buttonVariant: "invalid",
-          buttonText: t`Unknown Field`,
-          buttonIcon: (
-            <CloseIconButton
-              onClick={e => {
-                handleChangeTarget(null);
-                e.stopPropagation();
-              }}
-            />
-          ),
-        };
-      } else {
-        return {
-          buttonVariant: "default",
-          buttonTooltip: null,
-          buttonText: t`Select…`,
-          buttonIcon: <ChevrondownIcon name="chevrondown" />,
-        };
-      }
-    }, [
-      hasPermissionsToMap,
-      isDisabled,
-      selectedMappingOption,
-      target,
-      handleChangeTarget,
-      isVirtual,
-    ]);
 
   const layoutHeight = isMobile
     ? MOBILE_HEIGHT_BY_DISPLAY_TYPE[dashcard.card.display] ||
@@ -314,48 +210,17 @@ export function DashCardCardParameterMapper({
             </Header>
           )}
           <Flex align="center" justify="center" gap="xs" pos="relative">
-            <DeprecatedTooltip tooltip={buttonTooltip}>
-              <TippyPopover
-                visible={
-                  isDropdownVisible && !isDisabled && hasPermissionsToMap
-                }
-                onClickOutside={() => setIsDropdownVisible(false)}
-                placement="bottom-start"
-                content={
-                  <ParameterTargetList
-                    onChange={(target: ParameterTarget) => {
-                      handleChangeTarget(target);
-                      setIsDropdownVisible(false);
-                    }}
-                    target={target}
-                    mappingOptions={mappingOptions}
-                  />
-                }
-              >
-                <TargetButton
-                  variant={buttonVariant}
-                  aria-label={buttonTooltip ?? undefined}
-                  aria-haspopup="listbox"
-                  aria-expanded={isDropdownVisible}
-                  aria-disabled={isDisabled || !hasPermissionsToMap}
-                  onClick={() => {
-                    setIsDropdownVisible(true);
-                  }}
-                  onKeyDown={e => {
-                    if (e.key === "Enter") {
-                      setIsDropdownVisible(true);
-                    }
-                  }}
-                >
-                  {buttonText && (
-                    <TargetButtonText>
-                      <Ellipsified>{buttonText}</Ellipsified>
-                    </TargetButtonText>
-                  )}
-                  {buttonIcon}
-                </TargetButton>
-              </TippyPopover>
-            </DeprecatedTooltip>
+            <DashCardCardParameterMapperButton
+              handleChangeTarget={handleChangeTarget}
+              isVirtual={isVirtual}
+              isQuestion={isQuestion}
+              isDisabled={isDisabled}
+              selectedMappingOption={selectedMappingOption}
+              question={question}
+              card={card}
+              target={target}
+              mappingOptions={mappingOptions}
+            />
             {shouldShowAutoConnectHint &&
               layoutHeight <= 3 &&
               dashcard.size_x > 4 && <AutoConnectedAnimatedIcon />}
